+++ src/main/java/com/tsg/superherosighting/controllers/LocationController.java
@@ -22,6 +22,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 /**
  *
@@ -112,8 +113,14 @@ public class LocationController {
     }
 
     @GetMapping("deleteLocation/{id}")
-    public String deleteLocation(@PathVariable Integer id) {
-        superDao.removeLocation(id);
-        return "redirect:/locations";
+    @ResponseBody
+    public Location deleteLocation(@PathVariable Integer id) {
+        Location toRemove = superDao.getALocation(id);
+        if (toRemove != null) {
+            superDao.removeLocation(id);
+            return toRemove;
+        } else {
+            return null;
+        }
     }
 }
