+++ src/main/java/com/tsg/superherosighting/controllers/PowerController.java
@@ -20,6 +20,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 /**
  *
@@ -27,12 +28,12 @@ import org.springframework.web.bind.annotation.PostMapping;
  */
 @Controller
 public class PowerController {
-
+    
     @Autowired
     SuperDaoDBJdbcImpl superDao;
-
+    
     Set<ConstraintViolation<SuperPower>> violations = new HashSet<>();
-
+    
     @GetMapping("powers")
     public String displayPowers(Model model) {
         List<SuperPower> powers = superDao.getAllSuperpowers();
@@ -40,43 +41,43 @@ public class PowerController {
         model.addAttribute("errors", violations);
         return "powers";
     }
-
+    
     @PostMapping("addPower")
     public String addPower(HttpServletRequest request) {
         String name = request.getParameter("name");
-
+        
         SuperPower power = new SuperPower();
         power.setName(name);
-
+        
         Validator validate = Validation.buildDefaultValidatorFactory().getValidator();
         violations = validate.validate(power);
-
+        
         if (violations.isEmpty()) {
             superDao.addSuperpower(power);
         }
-
+        
         return "redirect:powers";
     }
-
+    
     @GetMapping("editPower")
     public String editPower(HttpServletRequest request, Model model) {
         int id = Integer.parseInt(request.getParameter("id"));
         SuperPower power = superDao.getASuperpower(id);
-
+        
         model.addAttribute("power", power);
         model.addAttribute("errors", violations);
         return "editPower";
     }
-
+    
     @PostMapping("editPower")
     public String performEditPower(HttpServletRequest request, Model model) {
         int id = Integer.parseInt(request.getParameter("id"));
         SuperPower power = superDao.getASuperpower(id);
         power.setName(request.getParameter("name"));
-
+        
         Validator validate = Validation.buildDefaultValidatorFactory().getValidator();
         violations = validate.validate(power);
-
+        
         if (violations.isEmpty()) {
             superDao.editSuperpower(power);
             return "redirect:powers";
@@ -87,11 +88,17 @@ public class PowerController {
             return "editPower";
         }
     }
-
+    
     @GetMapping("deletePower/{id}")
-    public String deletePower(@PathVariable Integer id) {
-        superDao.removeSuperpower(id);
-        return "redirect:/powers";
+    @ResponseBody
+    public SuperPower deletePower(@PathVariable Integer id) {
+        SuperPower toRemove = superDao.getASuperpower(id);
+        if (toRemove != null) {
+            superDao.removeSuperpower(id);
+            return toRemove;
+        } else {
+            return null;
+        }
     }
-
+    
 }
