+++ src/main/java/com/tsg/superherosighting/controllers/SightingController.java
@@ -8,7 +8,6 @@ package com.tsg.superherosighting.controllers;
 import com.tsg.superherosighting.dao.SuperDaoDBJdbcImpl;
 import com.tsg.superherosighting.dto.Hero;
 import com.tsg.superherosighting.dto.Location;
-import com.tsg.superherosighting.dto.Organization;
 import com.tsg.superherosighting.dto.Sighting;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -26,6 +25,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 /**
  *
@@ -42,11 +42,11 @@ public class SightingController {
     @GetMapping("sightings")
     public String displaySightings(Model model) {
         List<Hero> heroes = superDao.getAllHeroes();
-        
+
         List<Location> locations = superDao.getAllLocations();
         Location naLocation = superDao.getALocation(-1);
         locations.remove(naLocation);
-        
+
         List<Sighting> sightings = superDao.getAllSightings();
         model.addAttribute("currentTime", LocalDateTime.now().withNano(0).withSecond(0));
         model.addAttribute("heroes", heroes);
@@ -142,7 +142,7 @@ public class SightingController {
             List<Location> allLocs = superDao.getAllLocations();
             Location naLocation = superDao.getALocation(-1);
             allLocs.remove(naLocation);
-            
+
             model.addAttribute("currentTime", LocalDateTime.now().withNano(0).withSecond(0));
             model.addAttribute("heroes", allHeroes);
             model.addAttribute("locations", allLocs);
@@ -154,8 +154,14 @@ public class SightingController {
     }
 
     @GetMapping("deleteSighting/{id}")
-    public String deleteSighting(@PathVariable Integer id) {
-        superDao.removeSighting(id);
-        return "redirect:/sightings";
+    @ResponseBody
+    public Sighting deleteSighting(@PathVariable Integer id) {
+        Sighting toRemove = superDao.getASighting(id);
+        if (toRemove != null) {
+            superDao.removeSighting(id);
+            return toRemove;
+        } else {
+            return null;
+        }
     }
 }
